updated pointclouds for ones with bigger effectors

+++ modules/toolFeatExt/include/tool3DFeat_IDLServer.h
@@ -50,9 +50,10 @@ public:
   /**
    * @brief setCanonicalPose - Rotates the tool model to canonical orientations, i.e. as a rotation along the longest axis which orients the end effector.
    * @param deg - (double) degrees of rotation along the longest axis.
+   * @param disp - (int) displacement along the tool axis (grasped closer or further from end-effector).
    * @return true/false on success/failure of rotating model according to orientation.
    */
-  virtual bool setCanonicalPose(const double deg = 0);
+  virtual bool setCanonicalPose(const int32_t deg = 0);
   /**
    * @brief bins - sets the number of bins per angular dimension (yaw-pitch-roll) used to compute the normal histogram. Total number of bins per voxel = bins^3.
    * @param nbins - (int) desired number of bins per angular dimension. (default = 2, i.e. 8 bins per voxel).
